Revert to a41ae7a: roll back production to pre-feature-expansion state
Reverts every commit merged to main after a41ae7a8 ("fix: replace state
190/491 occupation lists with real published data", #44), in a single
commit for a clean rollback point:

  60eee5a program status badges, PR191 pathway, cost estimate, report
          export, occupation comparison (#45)
  8db4755 5-round audit — print regression, data corrections, type
          safety, i18n (#46)
  2ab37f1 employer sponsorship checklist (482/186) (#47)
  05262dc employer sponsorship section layout fix (#48)
  951a4aa split employer sponsorship onto its own /sponsorship page (#49)
  6c519e7 extract shared profile data onto its own page, unify page
          widths (#50)
  966d2e0 pathway-status messaging and reset button fixes (#51)
  72b9485 fixed-element overlaps and scroll affordance on mobile (#52)
  31a94b4 dead i18n keys and zh punctuation consistency (#53)
  bfd1d3f cross-tab persistence clobbering, lint, dead files (#54)

Requested rollback — production is going back to exactly a41ae7a8's
tree. Confirmed scope and mechanism (revert, not a history rewrite)
directly with the requester before applying.
