--- eslint.config.mjs
@@ -1,7 +1,0 @@
-import nextConfig from 'eslint-config-next';
-
-const config = [
-  ...nextConfig,
-];
-
-export default config;
