+++ package.json
@@ -6,7 +6,6 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "eslint .",
     "test": "vitest run",
     "test:watch": "vitest"
   },
