+++ src/app/i18n/index.ts
@@ -1,0 +1,23 @@
+import { createInstance } from 'i18next';
+import resourcesToBackend from 'i18next-resources-to-backend';
+import { initReactI18next } from 'react-i18next';
+import { getOptions } from './settings';
+
+const initI18next = async (lng: string, ns: string) => {
+  const i18nInstance = createInstance();
+  await i18nInstance
+    .use(initReactI18next)
+    .use(resourcesToBackend((language: string, namespace: string) => {
+      return import(`../../../public/locales/${language}/${namespace}.json`);
+    }))
+    .init(getOptions(lng, ns));
+  return i18nInstance;
+};
+
+export async function useTranslation(lng: string, ns: string, options: { keyPrefix?: string } = {}) {
+  const i18nextInstance = await initI18next(lng, ns);
+  return {
+    t: i18nextInstance.getFixedT(lng, ns, options.keyPrefix),
+    i18n: i18nextInstance,
+  };
+}
