+++ src/app/i18n/settings.ts
@@ -1,0 +1,14 @@
+export const fallbackLng = 'en';
+export const languages = ['en', 'zh'];
+export const defaultNS = 'common';
+
+export function getOptions(lng = fallbackLng, ns = defaultNS) {
+  return {
+    supportedLngs: languages,
+    fallbackLng,
+    lng,
+    fallbackNS: defaultNS,
+    defaultNS,
+    ns,
+  };
+}
