+++ src/components/ExportModal.tsx
@@ -1,14 +1,11 @@
 'use client';
 
 import { useEffect, useRef, useState } from 'react';
-import { createPortal } from 'react-dom';
 import { useTranslation } from 'react-i18next';
 import type { Evaluation } from '@/lib/points';
 import { drawCard, ensureCardFonts } from '@/lib/exportCard';
 import type { CardLabels } from '@/lib/exportCard';
 import type { CardTheme } from '@/data/cardThemes';
-import type { JobAssessment, PlanningDates, SharedCriteria } from '@/lib/types';
-import ReportView from './ReportView';
 import { todayLabel } from './Header';
 
 interface ExportModalProps {
@@ -16,13 +13,7 @@ interface ExportModalProps {
   onClose: () => void;
   evaluation: Evaluation;
   goal: number;
-  shared: SharedCriteria;
-  jobs: JobAssessment[];
-  dates: PlanningDates;
-  today: string;
 }
-
-type ExportMode = 'card' | 'report';
 
 const BD_KEYS = [
   'age', 'english', 'education', 'partnerStatus',
@@ -30,11 +21,10 @@ const BD_KEYS = [
   'ausWork', 'overseasWork', 'professionalYear',
 ];
 
-export default function ExportModal({ open, onClose, evaluation, goal, shared, jobs, dates, today }: ExportModalProps) {
+export default function ExportModal({ open, onClose, evaluation, goal }: ExportModalProps) {
   const { t, i18n } = useTranslation();
   const lang = i18n.language?.startsWith('zh') ? 'zh' : 'en';
 
-  const [mode, setMode] = useState<ExportMode>('card');
   const [cardTheme, setCardTheme] = useState<CardTheme>('cream');
   const [img, setImg] = useState('');
   const [loading, setLoading] = useState(false);
@@ -70,7 +60,7 @@ export default function ExportModal({ open, onClose, evaluation, goal, shared, j
   };
 
   useEffect(() => {
-    if (!open || mode !== 'card') return;
+    if (!open) return;
     let cancelled = false;
     const seq = (genSeqRef.current += 1);
     setLoading(true);
@@ -85,7 +75,6 @@ export default function ExportModal({ open, onClose, evaluation, goal, shared, j
         cardGoal: t('cardGoal'),
         cardMin: t('cardMin'),
         noBestPath: t('noBestPath'),
-        noPathEligible: t('noPathEligible'),
         bestPathPrefix: t('bestPathPrefix'),
         cardEmpty: t('cardEmpty'),
         cardShared: t('cardShared'),
@@ -106,7 +95,7 @@ export default function ExportModal({ open, onClose, evaluation, goal, shared, j
       }
     })();
     return () => { cancelled = true; };
-  }, [open, mode, cardTheme, lang, evaluation, goal, t]);
+  }, [open, cardTheme, lang, evaluation, goal, t]);
 
   if (!open) return null;
 
@@ -120,8 +109,6 @@ export default function ExportModal({ open, onClose, evaluation, goal, shared, j
     a.click();
   };
 
-  const printReport = () => window.print();
-
   const themeTab = (value: CardTheme, label: string) => (
     <button
       type="button"
@@ -133,23 +120,6 @@ export default function ExportModal({ open, onClose, evaluation, goal, shared, j
         border: 'none',
         color: cardTheme === value ? 'var(--overlay-ink)' : 'var(--overlay-dim)',
         borderBottom: cardTheme === value ? '1px solid var(--overlay-ink)' : '1px solid transparent',
-      }}
-    >
-      {label}
-    </button>
-  );
-
-  const modeTab = (value: ExportMode, label: string) => (
-    <button
-      type="button"
-      onClick={() => setMode(value)}
-      aria-pressed={mode === value}
-      className="cursor-pointer text-xs tracking-[0.1em] py-3 -my-2"
-      style={{
-        background: 'none',
-        border: 'none',
-        color: mode === value ? 'var(--overlay-ink)' : 'var(--overlay-dim)',
-        borderBottom: mode === value ? '1px solid var(--overlay-ink)' : '1px solid transparent',
       }}
     >
       {label}
@@ -175,15 +145,12 @@ export default function ExportModal({ open, onClose, evaluation, goal, shared, j
         tabIndex={-1}
         onKeyDown={trapTab}
         className="relative flex flex-col gap-3.5 outline-none"
-        style={{ width: mode === 'card' ? 'min(86vw, 380px)' : 'min(92vw, 640px)', animation: 'eoiModalIn 0.4s cubic-bezier(0.22, 1, 0.36, 1) backwards' }}
+        style={{ width: 'min(86vw, 380px)', animation: 'eoiModalIn 0.4s cubic-bezier(0.22, 1, 0.36, 1) backwards' }}
       >
         <div className="flex justify-between items-center">
           <div className="flex gap-3.5 text-xs">
-            {modeTab('card', t('exportModeCard'))}
-            {modeTab('report', t('exportModeReport'))}
-            {mode === 'card' && <span style={{ width: '1px', alignSelf: 'stretch', background: 'var(--overlay-hair)' }} />}
-            {mode === 'card' && themeTab('cream', t('cardCream'))}
-            {mode === 'card' && themeTab('charcoal', t('cardCharcoal'))}
+            {themeTab('cream', t('cardCream'))}
+            {themeTab('charcoal', t('cardCharcoal'))}
           </div>
           <button
             type="button"
@@ -195,59 +162,39 @@ export default function ExportModal({ open, onClose, evaluation, goal, shared, j
             ×
           </button>
         </div>
-        {mode === 'card' ? (
-          <div className="flex justify-center" style={{ maxHeight: '64vh' }}>
-            {img && !loading && (
-              /* Canvas-generated preview, not a static asset */
-              /* eslint-disable-next-line @next/next/no-img-element */
-              <img
-                src={img}
-                alt={t('cardPreviewAlt')}
-                className="block"
-                style={{ maxWidth: '100%', maxHeight: '64vh', boxShadow: 'var(--overlay-shadow)' }}
-              />
-            )}
-            {loading && (
-              <div
-                className="w-full flex items-center justify-center text-[0.78125rem] tracking-[0.12em]"
-                style={{
-                  aspectRatio: '3 / 4',
-                  background: 'var(--overlay-surface)',
-                  border: '1px solid var(--overlay-hair-soft)',
-                  color: 'var(--overlay-muted)',
-                }}
-              >
-                {t('generating')}
-              </div>
-            )}
-          </div>
-        ) : (
-          <div className="overflow-y-auto" style={{ maxHeight: '64vh', boxShadow: 'var(--overlay-shadow)' }}>
-            <ReportView evaluation={evaluation} shared={shared} jobs={jobs} goal={goal} dates={dates} today={today} dateLabel={todayLabel()} />
-          </div>
-        )}
-        {/*
-          On-screen preview above lives deep inside this fixed-position modal, which is fine for
-          display but hostile to printing: the modal sits on top of the whole (still-in-flow) page,
-          so hiding "everything except the report" by walking up the DOM either leaves stray blank
-          pages or (with visibility tricks) hides the report too — see globals.css history. Instead,
-          print from an independent copy portaled directly onto <body>, a sibling of <main> — trivial
-          to isolate with a single CSS rule and immune to whatever the rest of the page is doing.
-        */}
-        {mode === 'report' && typeof document !== 'undefined' && createPortal(
-          <div className="print-report">
-            <ReportView evaluation={evaluation} shared={shared} jobs={jobs} goal={goal} dates={dates} today={today} dateLabel={todayLabel()} />
-          </div>,
-          document.body,
-        )}
+        <div className="flex justify-center" style={{ maxHeight: '64vh' }}>
+          {img && !loading && (
+            /* Canvas-generated preview, not a static asset */
+            /* eslint-disable-next-line @next/next/no-img-element */
+            <img
+              src={img}
+              alt={t('cardPreviewAlt')}
+              className="block"
+              style={{ maxWidth: '100%', maxHeight: '64vh', boxShadow: 'var(--overlay-shadow)' }}
+            />
+          )}
+          {loading && (
+            <div
+              className="w-full flex items-center justify-center text-[0.78125rem] tracking-[0.12em]"
+              style={{
+                aspectRatio: '3 / 4',
+                background: 'var(--overlay-surface)',
+                border: '1px solid var(--overlay-hair-soft)',
+                color: 'var(--overlay-muted)',
+              }}
+            >
+              {t('generating')}
+            </div>
+          )}
+        </div>
         <div className="flex gap-2.5">
           <button
             type="button"
-            onClick={mode === 'card' ? download : printReport}
+            onClick={download}
             className="flex-1 cursor-pointer text-[0.78125rem] tracking-[0.18em] font-medium px-5 py-3.5 hover:opacity-88"
             style={{ background: 'var(--overlay-ink)', color: 'var(--overlay-contrast)', border: 'none' }}
           >
-            {mode === 'card' ? t('download') : t('printReport')}
+            {t('download')}
           </button>
           <button
             type="button"
