+++ src/components/FloatingChip.tsx
@@ -17,7 +17,7 @@ export default function FloatingChip({ visible, total, onClick }: FloatingChipPr
       type="button"
       onClick={onClick}
       title={t('chipHint')}
-      className="fixed right-[22px] z-40 flex items-baseline gap-2.5 px-[18px] py-[11px] cursor-pointer"
+      className="fixed right-[22px] z-40 flex items-baseline gap-2.5 px-[18px] py-[11px] cursor-pointer wide:hidden"
       style={{
         bottom: 'calc(22px + env(safe-area-inset-bottom, 0px))',
         background: 'var(--band-bg)',
