+++ src/components/TimelineChart.tsx
@@ -6,7 +6,6 @@ import { useTranslation } from 'react-i18next';
 import { addMonths, monthsBetween } from '@/lib/timeline';
 import type { TimelineResult } from '@/lib/timeline';
 import { MIN_POINTS } from '@/data/pointsCriteria';
-import HScrollFade from './HScrollFade';
 
 interface TimelineChartProps {
   timeline: TimelineResult;
@@ -148,7 +147,7 @@ function TimelineChart({ timeline, goal, today, focusEventIndex = null, seriesLa
           </span>
         ))}
       </div>
-      <HScrollFade>
+      <div className="overflow-x-auto">
       <svg
         viewBox={`0 0 ${W} ${H}`}
         role="img"
@@ -292,7 +291,7 @@ function TimelineChart({ timeline, goal, today, focusEventIndex = null, seriesLa
           </g>
         )}
       </svg>
-      </HScrollFade>
+      </div>
     </div>
   );
 }
