--- src/data/csol.ts
@@ -1,74 +1,0 @@
-// Core Skills Occupation List (CSOL) — gates the subclass 482 Core Skills
-// stream and (per the parallel legislative instrument) the subclass 186
-// Direct Entry stream. A DIFFERENT list from MLTSSL/STSOL/ROL (occupations.ts),
-// which gate the points-tested 189/190/491 pathways — overlap is substantial
-// but neither list is a subset of the other.
-// Legal basis: Migration (Specification of Occupations—Subclass 482 Visa)
-// Instrument 2024 (F2024L01620, LIN 24/089); 186 Direct Entry uses the
-// parallel F2024L01618 (LIN 24/093), same occupation schedule.
-// 456 occupations, cross-verified 2026-07 against the official PDF
-// (immi.homeaffairs.gov.au/Documents/core-sol.pdf) and an independent legal
-// reproduction (workvisalawyers.com.au) — both agree on the count and the
-// final entry (639211 Retail Buyer). Home Affairs' live search tool tags one
-// additional code (149914 Financial Institution Branch Manager) as CSOL-482
-// eligible that is absent from the official PDF; excluded here pending
-// confirmation in a future PDF revision.
-// Some listed occupations carry additional salary/duty caveats (e.g. CEO
-// nominations under AUD180,001 are excluded even though 111111 is listed) —
-// this file only captures list membership, not caveat text. Verify caveats
-// against the official list before relying on this for a real nomination.
-export const csolCodes: readonly string[] = [
-  "111111", "111211", "121111", "121311", "121313", "121315", "121318", "121321", "121611", "131112",
-  "131113", "132111", "132211", "132311", "132411", "132511", "133111", "133112", "133211", "133511",
-  "133512", "133611", "133612", "134211", "134212", "134213", "134311", "134411", "134499", "135111",
-  "135112", "135199", "139911", "139912", "139913", "139916", "139917", "141311", "141411", "141999",
-  "142111", "142116", "149411", "149911", "149912", "149915", "149999", "211212", "212111", "212315",
-  "212316", "212317", "212318", "212413", "212414", "212415", "212416", "212499", "221111", "221112",
-  "221113", "221211", "221213", "221214", "222112", "222113", "222311", "223111", "223112", "223113",
-  "224111", "224112", "224114", "224115", "224116", "224511", "224512", "224712", "224713", "224714",
-  "224914", "224999", "225111", "225113", "225114", "225211", "225212", "225213", "225311", "225411",
-  "225412", "225499", "231111", "231113", "231114", "231199", "231212", "232111", "232112", "232212",
-  "232213", "232214", "232313", "232412", "232413", "232414", "232511", "232611", "233111", "233112",
-  "233211", "233212", "233213", "233214", "233215", "233311", "233411", "233511", "233512", "233513",
-  "233611", "233612", "233911", "233912", "233913", "233914", "233915", "233916", "233999", "234111",
-  "234114", "234115", "234116", "234211", "234212", "234213", "234312", "234399", "234411", "234412",
-  "234413", "234511", "234513", "234515", "234516", "234521", "234522", "234599", "234612", "234711",
-  "234911", "234912", "234913", "234914", "234999", "241111", "241213", "241311", "241411", "241511",
-  "241512", "241513", "241599", "242111", "242211", "249112", "249214", "249299", "251111", "251211",
-  "251212", "251213", "251214", "251312", "251411", "251412", "251511", "251512", "251513", "251912",
-  "251999", "252214", "252299", "252311", "252312", "252411", "252511", "252611", "252711", "252712",
-  "253111", "253112", "253211", "253311", "253312", "253313", "253314", "253315", "253316", "253317",
-  "253318", "253321", "253322", "253323", "253324", "253399", "253411", "253511", "253512", "253513",
-  "253514", "253515", "253516", "253517", "253518", "253521", "253911", "253912", "253913", "253914",
-  "253915", "253917", "253918", "253999", "254111", "254211", "254212", "254411", "254412", "254413",
-  "254414", "254415", "254416", "254417", "254418", "254421", "254422", "254423", "254424", "254425",
-  "254499", "261111", "261112", "261211", "261212", "261311", "261312", "261313", "261314", "261315",
-  "261316", "261317", "261399", "262111", "262113", "262114", "262115", "262116", "262117", "262118",
-  "263111", "263112", "263113", "263211", "263213", "263299", "263312", "271111", "271214", "271299",
-  "271311", "272112", "272114", "272115", "272311", "272312", "272313", "272314", "272399", "272413",
-  "272511", "272612", "311112", "311113", "311114", "311115", "311211", "311212", "311215", "311217",
-  "311299", "311312", "311314", "311399", "311411", "311412", "311499", "312111", "312112", "312113",
-  "312114", "312116", "312199", "312211", "312212", "312311", "312312", "312412", "312511", "312512",
-  "312911", "312912", "312913", "312914", "312999", "313111", "313112", "313113", "313199", "313212",
-  "313213", "313214", "321111", "321211", "321212", "321213", "321214", "322112", "322113", "322114",
-  "322211", "322311", "322312", "322313", "323111", "323112", "323113", "323211", "323212", "323213",
-  "323214", "323215", "323299", "323313", "323314", "323411", "323412", "324111", "324211", "324212",
-  "324311", "331111", "331112", "331211", "331212", "331213", "332111", "332211", "333111", "333211",
-  "333212", "333311", "333411", "334112", "334113", "334114", "334115", "334116", "334117", "341111",
-  "341112", "342111", "342211", "342212", "342311", "342313", "342314", "342315", "342411", "342412",
-  "342413", "342414", "351111", "351112", "351211", "351311", "351411", "361111", "361112", "361311",
-  "362411", "362511", "362512", "362711", "362712", "391111", "392111", "392112", "392211", "392311",
-  "393114", "393311", "394112", "394113", "394211", "394212", "394213", "394299", "399111", "399112",
-  "399211", "399212", "399213", "399513", "399516", "399599", "399611", "399911", "399913", "399914",
-  "399916", "399918", "399999", "411111", "411112", "411211", "411212", "411213", "411214", "411311",
-  "411411", "411611", "411711", "411713", "411715", "411716", "421111", "421114", "431411", "451111",
-  "451412", "451612", "451711", "452311", "452317", "452321", "511111", "511112", "512111", "521212",
-  "599111", "599211", "599612", "599915", "611211", "639211",
-];
-
-const csolSet: ReadonlySet<string> = new Set(csolCodes);
-
-/** Whether an ANZSCO code is on the Core Skills Occupation List (482 Core Skills / 186 Direct Entry gate). */
-export function isCsolListed(anzsco: string): boolean {
-  return anzsco !== '' && csolSet.has(anzsco);
-}
