+++ src/lib/points.ts
@@ -140,43 +140,3 @@ export function evaluate(shared: SharedCriteria, jobs: JobAssessment[]): Evaluat
 
   return { shared: sharedPoints, sharedTotal, jobs: jobEvaluations, best, bestTotal, bareScore };
 }
-
-/** Narrows `occupation` from nullable to present — lets `.filter(hasOccupation).map(...)` read `je.occupation` without a `!` assertion. */
-export function hasOccupation(je: JobEvaluation): je is JobEvaluation & { occupation: Occupation } {
-  return je.occupation !== null;
-}
-
-/** The highest-scoring eligible pathway for a single assessment, or null if none is eligible. */
-export function bestPathwayForJob(je: JobEvaluation): PathwayResult | null {
-  let best: PathwayResult | null = null;
-  for (const p of je.pathways) {
-    if (p.eligible && (!best || p.total > best.total)) best = p;
-  }
-  return best;
-}
-
-export type PathwayStatus = 'noOcc' | 'listNo' | 'noState' | 'low' | 'ok';
-
-/**
- * Why a pathway is or isn't eligible, in the same precedence every ineligibility
- * display in the app should follow: no occupation picked, then the federal
- * list gate, then the per-state list gate (190/491 only), then the points
- * floor. Shared by ResultsBand (live UI) and ReportView (export) so the two
- * never drift into showing different reasons for the same pathway.
- */
-export function pathwayStatus(p: PathwayResult): PathwayStatus {
-  if (!p.hasOccupation) return 'noOcc';
-  if (!p.listOk) return 'listNo';
-  if (p.code !== '189' && p.states.length === 0) return 'noState';
-  if (p.total < MIN_POINTS) return 'low';
-  return 'ok';
-}
-
-/** i18n key per `pathwayStatus()` result — shared so every surface labels the same reason the same way. */
-export const PATHWAY_STATUS_LABEL_KEY: Record<PathwayStatus, string> = {
-  noOcc: 'pathNoOcc',
-  listNo: 'pathListNo',
-  noState: 'pathNoState',
-  low: 'pathLow',
-  ok: 'pathOk',
-};
