+++ src/lib/types.ts
@@ -30,11 +30,9 @@ export interface PlanningDates {
   birth: string;
   englishTest: string;
   naatiCert: string;
-  /** 491/494 visa grant month — powers the subclass 191 (permanent) eligibility projection */
-  visa491Grant: string;
 }
 
-export const defaultPlanningDates: PlanningDates = { birth: '', englishTest: '', naatiCert: '', visa491Grant: '' };
+export const defaultPlanningDates: PlanningDates = { birth: '', englishTest: '', naatiCert: '' };
 
 /** Month-string guard shared by url parsing, timeline math and inputs */
 export function isYm(s: string): boolean {
@@ -50,20 +48,6 @@ export const defaultSharedCriteria: SharedCriteria = {
   ausStudy: false,
   regionalStudy: false,
   communityLanguage: false,
-};
-
-/** Inputs for the employer-sponsorship checklist (482/186) — persisted separately from the points-tested state. */
-export interface SponsorshipInputs {
-  hasSponsor: boolean;
-  salaryBand: '' | 'belowCsit' | 'csitToSsit' | 'ssitPlus';
-  /** Cumulative ≥2 years on the nominating employer's 482/457 in the nominated occupation (186 TRT gate) */
-  trtEligible: boolean;
-}
-
-export const defaultSponsorshipInputs: SponsorshipInputs = {
-  hasSponsor: false,
-  salaryBand: '',
-  trtEligible: false,
 };
 
 let jobSeq = 0;
